More front and back end code updates
* Refactored the overall filesystem and integrity checker code. Checks
are now done in two main passes, the instant (startup) and the late
(main menu). This allows critical checks (like the x64 or the renderer
ones) to be completed before anything else forces KSP to crash (or
close). Some basic logging is also done in the startup pass for the same
reason.
* Added a check for the renderer type when using Windows. DirectX11 and
OpenGL do not work correctly with Scatterer when enforced via the
command line switches.
* Fixed some nullrefs caused by the PopupDialog.
* The missing dependencies will now be listed in a single notification
window instead of spamming the main menu screen with separate windows.
* Simplified the overall code wherever applicable.